Implemented Security Classifications and Redaction [#3 responsible:none state:fixed]
Users can now set a security classification on specific records so that only
those users with the classification (or a higher classification) can view the
record. If the user does not have the necessary classification, the record is
omitted from all SQL result sets.

The five default security classification (Unclassified, Restricted,
Confidential, Secret and Top Secret) are defined as security roles. Additional
speciality classifications can also be set up (as security roles) to provide
fine grained access control to specific records.

Redaction functionality is also available to obscure specific fields on a
record. This allows a user can view most of the record information
without revealing sensitive or confidential details (e.g. the amount for
a mortgage or the rental for a lease, etc).